More cleanups, the old DttSP is now call compatible again with the new one.

+++ trunk/Software/DttSP/update.cpp
@@ -2157,7 +2157,7 @@ SetDeviation (float value)
 }
 
 void
-SetRIT (float freq)
+SetRIT (double freq)
 {
 }
